fix: default blocks to 20000, reset pools if set empty
